% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funcModules}
\alias{funcModules}
\title{7,763 functional clusters}
\format{list of functional clusters
\describe{
  \item{each item}{list of function/phenotype terms of given cluster}
  ...
}}
\usage{
funcModules
}
\description{
A dataset containing the functional clusters
}
\keyword{datasets}
