% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gutTaxo}
\alias{gutTaxo}
\title{MSP taxonomy table}
\format{data frame with taxonomy
\describe{
  \item{MSP}{MSP id}
  \item{size}{the number of genes included}
  \item{species}{species-level}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
gutTaxo
}
\description{
A dataset containing the taxonomy of MSPs
}
\keyword{datasets}
