% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gutTaxoTab}
\alias{gutTaxoTab}
\title{gut MSP taxonomy table}
\format{data frame with taxonomy
\describe{
  \item{MSP}{MSP id}
  \item{size}{the number of genes included}
  \item{species}{species-level}
  ...
}}
\source{
\url{http://www.diamondse.info/}

\url{http://www.diamondse.info/}
}
\usage{
gutTaxoTab

gutTaxoTab
}
\description{
A dataset containing the taxonomy of MSPs

A dataset containing the functional clusters
}
\keyword{datasets}
