% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koDescMap}
\alias{koDescMap}
\title{KEGG orthology description table}
\format{matrix
\describe{
  \item{KO}{KO ID}
  \item{gene}{Gene ID}
  \item{desc}{KO description}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
koDescMap
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
