% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metaTab}
\alias{metaTab}
\title{longitudinal Swedish cohort metadata}
\format{data frame
\describe{
  \item{sampleId}{each id corresponds to the column of mgsMat}
  \item{metadataId}{each subject id}
  \item{time}{time passed after first visit (days)}
}}
\usage{
metaTab
}
\description{
A dataset containing sample metadata of longitudinal Swedish cohort
}
\keyword{datasets}
