% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgsMat}
\alias{mgsMat}
\title{MSP abundance matrix of longitudinal Swedish cohort}
\format{matrix
\describe{
  \item{row}{MSPs}
  \item{column}{KEGG orthology terms}
}}
\usage{
mgsMat
}
\description{
A dataset containing MSP abundance profile
}
\keyword{datasets}
