% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koDescMap}
\alias{koDescMap}
\title{KEGG orthology description table}
\format{data frame
\describe{
  \item{KO}{KO ID}
  \item{gene}{Gene ID}
  \item{desc}{KO description}
}}
\usage{
koDescMap
}
\description{
A dataset containing KEGG orthology term descriptions
}
\keyword{datasets}
