% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mspKoMat}
\alias{mspKoMat}
\title{KO-MSP mapping matrix}
\format{matrix
\describe{
  \item{row}{MSPs}
  \item{column}{KEGG orthology terms}
}}
\usage{
mspKoMat
}
\description{
A dataset containing KEGG orthology (KO) mapping to gut MSP
}
\keyword{datasets}
